organize taxon_filter unit tests better and remove legacy fastq entry points and tests
